% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.data.R
\name{add.data}
\alias{add.data}
\title{Add data to the package 'outbreaks'}
\usage{
add.data(x, dir = getwd())
}
\arguments{
\item{x}{a dataset to be added}

\item{dir}{the directory where the files should be saved}
}
\description{
This function take an outbreak data as input and outputs two files: i) this object, saved as a
compressed '.RData', and ii) a R file containing a template of roxygen2 documentation.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com}).
}

