% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_fais_2011.R
\docType{data}
\name{dengue_fais_2011}
\alias{dengue_fais_2011}
\title{Dengue on the island of Fais, Micronesia, 2011 (new version)}
\format{
A data frame with 57 rows and 2 columns
\describe{
\item{date_of_onset}{First day of the onset week}
\item{incidence}{Incidence is given as the number of new cases reported in the week beginning at \code{date_of_onset}}
}
}
\source{
Funk et al. (2016)
}
\usage{
dengue_fais_2011
}
\description{
These data describe incidence of clinical cases of Dengue
on the island of Fais, Micronesia.
}
\details{
The data on Dengue incidence reported by Funk et al. (2016) cover the period
from 2011-09-15 to 2012-02-14, over which time a total of 157 clinical cases
were reported among 294 residents. The first reported case is thought to be
the index case. The population of Fais is concentrated in a single population
centre, and is thought to have been immunologically naive at the time of
infection.
\subsection{Note}{

This dataset is the new version. If you want to use the old one for compatibility reason,
you can load it manually \code{load('data-raw/dep_dengue_fais_2011.rda')} and use it \code{dep_dengue_fais_2011}.
}
}
\section{Licence}{
\href{https://creativecommons.org/licenses/by/4.0/}{CC BY}
}

\examples{
## show first few weeks of Dengue incidence
head(dengue_fais_2011)

}
\references{
S. Funk, et al. 2016. Comparative Analysis of Dengue and Zika Outbreaks Reveals
Differences by Setting and Virus. PLOS Neglected Tropical Diseases, 10(12),
e0005173. \url{http://doi.org/10.1371/journal.pntd.0005173}
}
\seealso{
Other dengue: 
\code{\link{dengue_yap_2011}},
\code{\link{zika_yap_2007}}
}
\author{
\itemize{
\item Data from Funk et al. (2016), provided by \href{https://github.com/sbnfunk}{Sebastian Funk}.
\item Transfer to R and documentation by Finlay Campbell (\email{finlaycampbell93@gmail.com}).
\item Refactoring by \href{https://github.com/romainx}{Romain}.
}
}
\concept{dengue}
\keyword{datasets}
