% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_fais_2011.R
\docType{data}
\name{dengue_fais_2011_td}
\alias{dengue_fais_2011_td}
\title{Dengue on the island of Fais, Micronesia, 2011 (new version)}
\format{
{
A data frame with 57 rows and 3 columns
\describe{
\item{date_of_onset}{Date}
\item{incidence}{Number of cases}
}

}
}
\source{
Funk et al. (2016)
}
\usage{
dengue_fais_2011_td
}
\description{
These data describe incidence of clinical cases of Dengue
on the island of Fais, Micronesia.
}
\details{
The data on Dengue incidence reported by Funk et al. (2016) cover the period
from 2011-09-15 to 2012-02-14, over which time a total of 157 clinical cases
were reported among 294 residents. The first reported case is thought to be
the index case. The population of Fais is concentrated in a single population
centre, and is thought to have been immunologically naive at the time of
infection.

Note: this dataset is the new version. If you want to use the old one for compatibility reason,
please use instead \code{\link{dengue_fais_2011}}.
}
\examples{
## show first few weeks of Dengue incidence
head(dengue_fais_2011_td)

}
\references{
S. Funk, et al. 2016. Comparative Analysis of Dengue and Zika Outbreaks Reveals
Differences by Setting and Virus. PLOS Neglected Tropical Diseases, 10(12),
e0005173. \link{http://doi.org/10.1371/journal.pntd.0005173}
}
\author{
Data from Funk et al. (2016), provided by Sebastian Funk (github.com/sbnfunk).
Transfer to R and documentation by Finlay Campbell (\email{finlaycampbell93@gmail.com}).
}
\keyword{datasets}
