% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_yap_2011.R
\name{dengue_yap_2011_as_ts}
\alias{dengue_yap_2011_as_ts}
\title{Coerce the dataset to a \code{tsibble} object}
\usage{
dengue_yap_2011_as_ts()
}
\value{
The dataset as a \code{tsibble} if the \code{tsibble} is installed, else the
dataset
}
\description{
Ensure it is a valid timeseries (regular, ordered,
no duplication, no gaps)
}
\examples{
dengue_yap_2011_as_ts()
}
