% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_yap_2011.R
\docType{data}
\name{dengue_yap_2011_td}
\alias{dengue_yap_2011_td}
\title{Dengue on the Yap Main Islands, Micronesia, 2011}
\format{
A data frame with 185 rows and 3 columns
\describe{
\item{date_of_onset}{First day of the onset week}
\item{incidence}{Incidence is given as the number of new cases reported in the week beginning at \code{date_of_onset}}
}
}
\source{
Funk et al. (2016)
}
\usage{
dengue_yap_2011_td
}
\description{
These data describe incidence of clinical cases of Dengue
on the Yap Main Islands, Micronesia.
}
\details{
The data on Dengue incidence reported by Funk et al. (2016) cover the period
from 2011-07-07 to 2012-04-12, over which time a total of 978 cases were
reported among 7391 residents. Suspected Dengue cases were identified by the
Yap Department of Health, using the WHO (2009) case definition. A small
proportion of cases (9\%) were reported on outer islands and included in the
time series for the Yap Main Islands. Dengue virus serotype 2 was confirmed by
reverse transcriptase polymerase chain reaction by the CDC Dengue Branch,
Puerto Rico.
\subsection{Note}{

This dataset is the new version. If you want to use the old one for compatibility reason,
please use instead \code{\link{dengue_yap_2011_td}}.
}
}
\section{Licence}{
\href{https://creativecommons.org/licenses/by/4.0/}{CC BY}
}

\examples{
# show first few weeks of Dengue incidence
head(dengue_yap_2011_td)

}
\references{
S. Funk, et al. 2016. Comparative Analysis of Dengue and Zika Outbreaks Reveals
Differences by Setting and Virus. PLOS Neglected Tropical Diseases, 10(12),
e0005173. \url{http://doi.org/10.1371/journal.pntd.0005173}
}
\author{
\itemize{
\item Data from Funk et al. (2016), provided by Sebastian Funk (github.com/sbnfunk).
\item Transfer to R and documentation by Finlay Campbell (\email{finlaycampbell93@gmail.com}).
\item Refactoring by \href{https://github.com/romainx}{Romain}.
}
}
\keyword{datasets}
