% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_kikwit_1995.R
\name{ebola_kikwit_1995_as_ts}
\alias{ebola_kikwit_1995_as_ts}
\title{Coerce the dataset to a \code{tsibble} object}
\usage{
ebola_kikwit_1995_as_ts()
}
\value{
The dataset as a \code{tsibble} if the \code{tsibble} is installed, else the
dataset
}
\description{
Ensure it is a valid timeseries (regular, ordered,
no duplication, no gaps)
}
\examples{
ebola_kikwit_1995_as_ts()
}
