% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_sierraleone_2014.R
\name{ebola_sierraleone_2014_as_ts}
\alias{ebola_sierraleone_2014_as_ts}
\title{Coerce the dataset to a \code{tsibble} object}
\usage{
ebola_sierraleone_2014_as_ts()
}
\value{
The dataset as a \code{tsibble} if the \code{tsibble} is installed, else the
dataset
}
\description{
Ensure it is a valid timeseries (regular, ordered,
no duplication, no gaps)
}
\examples{
\dontrun{
ebola_sierraleone_2014_as_ts() \%>\%
group_by_key() \%>\%
index_by(year_month = ~ yearmonth(.)) \%>\%
summarise(incidence = sum(incidence)) \%>\%
ggplot(aes(x= year_month, y=incidence)) +
geom_line(aes(color=status)) +
geom_point
}
}
