% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluH7N9_china_2013.R
\docType{data}
\name{fluH7N9_china_2013}
\alias{fluH7N9_china_2013}
\title{Influenza A H7N9 in China, 2013}
\format{
A data frame with 136 rows and 8 columns
}
\source{
\url{https://datadryad.org/resource/doi:10.5061/dryad.2g43n/1/}
}
\usage{
fluH7N9_china_2013
}
\description{
These data comprise of 136 cases of influenza A H7N9 in China, analysed by Kucharski et al. (2014).
}
\examples{
## show first few cases
head(fluH7N9_china_2013)

}
\references{
A. Kucharski, H. Mills, A. Pinsent, C. Fraser, M. Van Kerkhove, C. A. Donnelly, and S. Riley. 2014.
Distinguishing between reservoir exposure and human-to-human transmission for emerging pathogens using case onset data.
PLOS Currents Outbreaks. Mar 7, edition 1. doi: 10.1371/currents.outbreaks.e1473d9bfc99d080ca242139a06c455f.

A. Kucharski, H. Mills, A. Pinsent, C. Fraser, M. Van Kerkhove, C. A. Donnelly, and S. Riley. 2014.
Data from: Distinguishing between reservoir exposure and human-to-human transmission for emerging pathogens using case onset data.
Dryad Digital Repository. http://dx.doi.org/10.5061/dryad.2g43n.
}
\author{
Data collated by Adam Kucharski et al. from ProMed, WHO, FluTrackers, news reports and research articles.
Transfer to R and documentation by Simon Frost (\email{sdwfrost@gmail.com}).
}
\keyword{datasets}
