% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluH7N9_china_2013.R
\name{fluH7N9_china_2013_as_ts}
\alias{fluH7N9_china_2013_as_ts}
\title{Coerce the dataset to a \code{tsibble} object}
\usage{
fluH7N9_china_2013_as_ts()
}
\value{
The dataset as a \code{tsibble} if the \code{tsibble} is installed, else the
dataset
}
\description{
Ensure it is a valid timeseries (regular, ordered,
no duplication, no gaps)
}
\examples{
\dontrun{
fluH7N9_china_2013_as_ts() \%>\%
group_by_key() \%>\%
index_by(date = ~ yearweek(.)) \%>\%
summarise(incidence = sum(incidence)) \%>\%
ggplot(aes(x= date, y=incidence, color=outcome)) +
geom_line() +
geom_point()
}
}
