% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaks-package.R
\name{legacy_mode}
\alias{legacy_mode}
\title{Load legacy version (v1.x) of datasets.}
\usage{
legacy_mode(envdir = parent.frame(), verbose = TRUE)
}
\arguments{
\item{envdir}{the environment where the data should be loaded.}

\item{verbose}{should item names be printed during loading?}
}
\description{
Load legacy (previous) version of all the datasets for compatibility reason.
The environment where data will be loaded can be chosen.
}
\examples{
\dontrun{
# Default version
head(dengue_fais_2011)

# A tibble: 6 x 2
#  date_of_onset incidence
#  <date>            <dbl>
# 1 2011-09-15            0
# 2 2011-09-22            0
# 3 2011-09-29            0

# Activate the compatibility mode
legacy_mode()

# Loading objects:
#  dengue_fais_2011
# Warning message:
# This function replaces datasets with the previous (legacy) version for compatibility reason

# Back to legacy (previous) version
head(dengue_fais_2011)

# A tibble: 6 x 3
#  onset_date    nr value
#  <date>     <dbl> <dbl>
# 1 2011-09-15     7     0
# 2 2011-09-22    14     0
# 3 2011-09-29    21     0
}
}
