% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salmonella.R
\docType{data}
\name{s_enteritidis_pt59}
\alias{s_enteritidis_pt59}
\title{Salmonella Enteritidis PT59 outbreak}
\format{An object of class \code{list} of length 2.}
\source{
Multi-agency Salmonella Enteritidis PT59 Outbreak Control Team,
chaired by Public Health England
}
\usage{
s_enteritidis_pt59
}
\description{
This dataset covers a food-borne outbreak of Salmonella Enteritidis PT59
investigated by Public Health England. The data includes a distribution
network, and genetic clusters of bacteria isolated in contaminated
patients. The data is anonymised: identifiers of the food distribution
network have been hashed. The object \code{s_enteritidis_pt59} is a list
containing:
}
\details{
\itemize{

\item \code{$graph}: a \code{data.frame} containing with two columns
\code{from} and \code{to} specifying (directed) edges of the food
distribution network, alongside reported cases (terminal branches, or 'tips').

\item \code{$cluster}: a \code{factor} indicating the genetic cluster of
named tips.

}
}
\author{
Thomas Inns \email{Thomas.Inns@phe.gov.uk}, Philip Ashton, Tim
Dallman, Roberto Vivancos. Hashing and port to R by Thibaut Jombart.
}
\keyword{datasets}
