% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zika_yap_2007.R
\name{zika_yap_2007_as_ts}
\alias{zika_yap_2007_as_ts}
\title{Coerce the dataset to a \code{tsibble} object}
\usage{
zika_yap_2007_as_ts()
}
\value{
The dataset as a \code{tsibble} if the \code{tsibble} is installed, else the
dataset
}
\description{
Ensure it is a valid timeseries (regular, ordered,
no duplication, no gaps)
}
\examples{
zika_yap_2007_as_ts()
}
