% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zika_yap_2007.R
\docType{data}
\name{zika_yap_2007}
\alias{zika_yap_2007}
\title{Zika on the Yap Main Islands, Micronesia, 2007 (new version)}
\format{
A data frame with 29 rows and 2 columns
\describe{
\item{date_of_onset}{First day of the onset week}
\item{incidence}{Incidence is given as the number of new cases reported in the week beginning at \code{date_of_onset}}
}
}
\source{
Funk et al. (2016)
}
\usage{
zika_yap_2007
}
\description{
These data describe weekly incidence of probable and confirmed cases of Zika virus
on the Yap Main Islands, Micronesia.
}
\details{
The data on weekly cases reported by Funk et al. (2016) cover the period from
2007-02-18 to 2007-09-02, over which time there were a total of 108 cases
classified as probable (59) or confirmed (49) in a population of 7391. Cases
were identified by a combination of prospective and retrospective
surveillance at all health centres on Yap.
}
\section{Licence}{
\href{https://creativecommons.org/licenses/by/4.0/}{CC BY}
}

\examples{
## show first few weeks of Zika incidence
head(zika_yap_2007)

}
\references{
S. Funk, et al. 2016. Comparative Analysis of Dengue and Zika Outbreaks Reveals
Differences by Setting and Virus. PLOS Neglected Tropical Diseases, 10(12),
e0005173. \url{http://doi.org/10.1371/journal.pntd.0005173}
}
\seealso{
Other dengue: 
\code{\link{dengue_fais_2011}},
\code{\link{dengue_yap_2011}}
}
\author{
\itemize{
\item Data from Funk et al. (2016), provided by \href{https://github.com/sbnfunk}{Sebastian Funk}.
\item Transfer to R and documentation by Finlay Campbell (\email{finlaycampbell93@gmail.com}).
\item Refactoring by \href{https://github.com/romainx}{Romain}.
}
}
\concept{dengue}
\keyword{datasets}
